formatted and extracted out to seperate component folder
